--- apps/client/src/pages/time-table/components/info/delete-button.css.ts
@@ -1,18 +1,0 @@
-import { recipe } from '@vanilla-extract/recipes';
-
-export const closeBtn = recipe({
-  base: {
-    position: 'absolute',
-    top: '0rem',
-    right: '1rem',
-    cursor: 'pointer',
-  },
-  variants: {
-    isFestivalDeleteMode: {
-      true: {},
-      false: {
-        display: 'none',
-      },
-    },
-  },
-});
