+++ apps/client/src/pages/time-table/components/time-table-board/time-table-board.tsx
@@ -1,11 +1,9 @@
-import { useEffect, useState } from 'react';
 import BoothOpenBox from '@pages/time-table/components/booth-open-box/booth-open-box';
 import Stage from '@pages/time-table/components/stage/stage';
 import TimeCell from '@pages/time-table/components/time-cell/time-cell';
 import TimeTableItem from '@pages/time-table/components/time-table-item/time-table-item';
 import { END_HOUR, HALF_HOUR_TO_MINUTES } from '@pages/time-table/constants';
 import { useImageDownload } from '@pages/time-table/hooks/use-image-download';
-import { usePatchTimeTableMutation } from '@pages/time-table/hooks/use-patch-time-table-mutation';
 import { TimeTableInfo } from '@pages/time-table/types/time-table-info-type';
 import { generateTableRow, parseTimeString } from '@pages/time-table/utils';
 
@@ -27,8 +25,6 @@ const TimeTableBoard = ({
   timeTableInfo,
   isEditTimeTableMode,
   isFestivalDeleteMode,
-  isComplete,
-  onToggleComplete,
 }: Props) => {
   const { elementRef, downloadImage } = useImageDownload<HTMLDivElement>({
     fileName: `${clickedFestivalTitle}`,
@@ -37,36 +33,6 @@ const TimeTableBoard = ({
   const isHalfHourOpen = Number(openMin) === HALF_HOUR_TO_MINUTES;
   const ticketOpenHour = isHalfHourOpen ? openHour + 1 : openHour;
   const cellNumber = generateTableRow(ticketOpenHour);
-
-  const [selectedItems, setSelectedItems] = useState<
-    { userTimetableId: number; isSelected: boolean }[]
-  >([]);
-  const mutate = usePatchTimeTableMutation();
-
-  const handleItemClick = (userTimetableId: number, isSelected: boolean) => {
-    setSelectedItems((prev) => {
-      const existingItem = prev.find(
-        (item) => item.userTimetableId === userTimetableId,
-      );
-
-      if (existingItem) {
-        return prev.filter((item) => item.userTimetableId !== userTimetableId);
-      } else {
-        return [...prev, { userTimetableId, isSelected }];
-      }
-    });
-  };
-
-  useEffect(() => {
-    if (isComplete) {
-      onToggleComplete();
-      const userTimetables = selectedItems.map((item) => ({
-        userTimetableId: item.userTimetableId,
-        isSelected: item.isSelected,
-      }));
-      mutate.mutate(userTimetables);
-    }
-  }, [isComplete, onToggleComplete]);
 
   return (
     <section className={styles.container} ref={elementRef}>
@@ -95,7 +61,7 @@ const TimeTableBoard = ({
                   stageCount={timeTableInfo.stageCount}
                   stageOrder={stage.stageOrder}
                   isEditTimeTableMode={isEditTimeTableMode}
-                  onClick={handleItemClick}
+                  onClick={() => console.log('')}
                 />
               ))}
             </div>
