+++ apps/client/src/shared/router/lazy.ts
@@ -54,7 +54,7 @@ export const AddFestivalPage = lazy(
   () => import('@pages/time-table/page/add-festival'),
 );
 export const TimeTablePage = lazy(
-  () => import('@pages/time-table/page/time-table'),
+  () => import('@pages/time-table/page/time-table-page'),
 );
 export const TimeTableLayout = lazy(
   () => import('@pages/time-table/page/time-table-layout'),
